+++ ui/actions/mute-rules/index.ts
@@ -1,5 +1,4 @@
 export {
-  bulkDeleteMuteRules,
   createMuteRule,
   deleteMuteRule,
   getMuteRule,